Pushed discord to main.py
